WIP actuators

+++ app/devices/Actuator/ActuatorPwm.cpp
@@ -2,7 +2,7 @@
 #include "ActuatorPwm.h"
 #include "Ticks.h"
 
-ActuatorPwm::ActuatorPwm(Actuator* _target, uint8_t _period) : target(_target) {
+ActuatorPwm::ActuatorPwm(Actuator* _target, uint16_t _period) : target(_target) {
     this->periodStartTime = 0;
     this->target->setActive(false);
     this->setPeriod(_period);
